Refactor to create Internal module, rename Quantity' to Quantity. Resolves #98 and #100.
