+++ dimensional.cabal
@@ -62,7 +62,8 @@ library
                        Numeric.Units.Dimensional.UnitNames,
                        Numeric.Units.Dimensional.UnitNames.InterchangeNames,
                        Numeric.Units.Dimensional.Variants
-  other-modules:       Numeric.Units.Dimensional.UnitNames.Internal
+  other-modules:       Numeric.Units.Dimensional.Internal,
+                       Numeric.Units.Dimensional.UnitNames.Internal
 
 test-suite tests
   type:                exitcode-stdio-1.0
